// frontend/src/app/api/models/autorizacion-vigente.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface AutorizacionVigente {
  otorgada_en?: (string | null);
  resumen_finalidad: string;
  url_politica: '/politica-de-tratamiento-de-datos';
  version_otorgada?: (string | null);
  version_politica: string;
  vigente_desde: string;
}