// frontend/src/app/api/models/contenido-ref-panel.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { EstadoEditorial } from '../models/estado-editorial';
import { Id } from '../models/id';
import { TipoEntidadContenido } from '../models/tipo-entidad-contenido';
export interface ContenidoRefPanel {
  estado_editorial: EstadoEditorial;
  id: Id;
  slug?: (string | null);
  tipo: TipoEntidadContenido;
  titulo: string;
}