// frontend/src/app/api/models/contenido-resumen-panel.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ActorRef } from '../models/actor-ref';
import { EstadoEditorial } from '../models/estado-editorial';
import { Id } from '../models/id';
import { TipoEntidadContenido } from '../models/tipo-entidad-contenido';
export interface ContenidoResumenPanel {
  actualizado_en: string;
  actualizado_por: ActorRef;
  estado_editorial: EstadoEditorial;
  fecha_ultima_revision?: (string | null);
  id: Id;
  publicado: boolean;
  slug?: (string | null);
  tipo: TipoEntidadContenido;
  titulo: string;
  url_publica?: (string | null);
  version: number;
}